+++ src/pocketmine/level/format/io/region/Anvil.php
@@ -23,67 +23,13 @@ declare(strict_types=1);
 
 namespace pocketmine\level\format\io\region;
 
-use pocketmine\level\format\Chunk;
-use pocketmine\level\format\ChunkException;
 use pocketmine\level\format\io\ChunkUtils;
 use pocketmine\level\format\SubChunk;
-use pocketmine\nbt\BigEndianNBTStream;
-use pocketmine\nbt\NBT;
 use pocketmine\nbt\tag\ByteArrayTag;
 use pocketmine\nbt\tag\CompoundTag;
-use pocketmine\nbt\tag\IntArrayTag;
-use pocketmine\nbt\tag\ListTag;
 
-class Anvil extends McRegion{
-
-	public const REGION_FILE_EXTENSION = "mca";
-
-	protected function nbtSerialize(Chunk $chunk) : string{
-		$nbt = new CompoundTag("Level", []);
-		$nbt->setInt("xPos", $chunk->getX());
-		$nbt->setInt("zPos", $chunk->getZ());
-
-		$nbt->setByte("V", 1);
-		$nbt->setLong("LastUpdate", 0); //TODO
-		$nbt->setLong("InhabitedTime", 0); //TODO
-		$nbt->setByte("TerrainPopulated", $chunk->isPopulated() ? 1 : 0);
-		$nbt->setByte("LightPopulated", $chunk->isLightPopulated() ? 1 : 0);
-
-		$subChunks = [];
-		foreach($chunk->getSubChunks() as $y => $subChunk){
-			if($subChunk->isEmpty()){
-				continue;
-			}
-
-			$tag = $this->serializeSubChunk($subChunk);
-			$tag->setByte("Y", $y);
-			$subChunks[] = $tag;
-		}
-		$nbt->setTag(new ListTag("Sections", $subChunks, NBT::TAG_Compound));
-
-		$nbt->setByteArray("Biomes", $chunk->getBiomeIdArray());
-		$nbt->setIntArray("HeightMap", $chunk->getHeightMapArray());
-
-		$entities = [];
-
-		foreach($chunk->getSavableEntities() as $entity){
-			$entities[] = $entity->saveNBT();
-		}
-
-		$nbt->setTag(new ListTag("Entities", $entities, NBT::TAG_Compound));
-
-		$tiles = [];
-		foreach($chunk->getTiles() as $tile){
-			$tiles[] = $tile->saveNBT();
-		}
-
-		$nbt->setTag(new ListTag("TileEntities", $tiles, NBT::TAG_Compound));
-
-		//TODO: TileTicks
-
-		$writer = new BigEndianNBTStream();
-		return $writer->writeCompressed(new CompoundTag("", [$nbt]), ZLIB_ENCODING_DEFLATE, RegionLoader::$COMPRESSION_LEVEL);
-	}
+class Anvil extends RegionLevelProvider{
+	use LegacyAnvilChunkTrait;
 
 	protected function serializeSubChunk(SubChunk $subChunk) : CompoundTag{
 		return new CompoundTag("", [
@@ -92,44 +38,6 @@ class Anvil extends McRegion{
 			new ByteArrayTag("SkyLight", ChunkUtils::reorderNibbleArray($subChunk->getBlockSkyLightArray(), "\xff")),
 			new ByteArrayTag("BlockLight", ChunkUtils::reorderNibbleArray($subChunk->getBlockLightArray()))
 		]);
-	}
-
-	protected function nbtDeserialize(string $data) : Chunk{
-		$nbt = new BigEndianNBTStream();
-		$chunk = $nbt->readCompressed($data);
-		if(!($chunk instanceof CompoundTag) or !$chunk->hasTag("Level")){
-			throw new ChunkException("Invalid NBT format");
-		}
-
-		$chunk = $chunk->getCompoundTag("Level");
-
-		$subChunks = [];
-		$subChunksTag = $chunk->getListTag("Sections") ?? [];
-		foreach($subChunksTag as $subChunk){
-			if($subChunk instanceof CompoundTag){
-				$subChunks[$subChunk->getByte("Y")] = $this->deserializeSubChunk($subChunk);
-			}
-		}
-
-		if($chunk->hasTag("BiomeColors", IntArrayTag::class)){
-			$biomeIds = ChunkUtils::convertBiomeColors($chunk->getIntArray("BiomeColors")); //Convert back to original format
-		}else{
-			$biomeIds = $chunk->getByteArray("Biomes", "", true);
-		}
-
-		$result = new Chunk(
-			$chunk->getInt("xPos"),
-			$chunk->getInt("zPos"),
-			$subChunks,
-			$chunk->hasTag("Entities", ListTag::class) ? $chunk->getListTag("Entities")->getValue() : [],
-			$chunk->hasTag("TileEntities", ListTag::class) ? $chunk->getListTag("TileEntities")->getValue() : [],
-			$biomeIds,
-			$chunk->getIntArray("HeightMap", [])
-		);
-		$result->setLightPopulated($chunk->getByte("LightPopulated", 0) !== 0);
-		$result->setPopulated($chunk->getByte("TerrainPopulated", 0) !== 0);
-		$result->setGenerated();
-		return $result;
 	}
 
 	protected function deserializeSubChunk(CompoundTag $subChunk) : SubChunk{
@@ -145,8 +53,12 @@ class Anvil extends McRegion{
 		return "anvil";
 	}
 
-	public static function getPcWorldFormatVersion() : int{
-		return 19133; //anvil
+	protected static function getRegionFileExtension() : string{
+		return "mca";
+	}
+
+	protected static function getPcWorldFormatVersion() : int{
+		return 19133;
 	}
 
 	public function getWorldHeight() : int{
